requirements.txt
streamlit
numpy
pandas
psycopg2-binary
streamlit-pills